completed moving circuit_latex.py
